+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.4.1
 
 - Fix RGB colour picker setting wrong colour on Hue lights (delay saturation write after hue)
+- Fix camera talk button not working (#114)
 
 ## 2.4.0